fix(cli): moved test to the correct directory
