+++ tooling/cli/test/extract-component-types.test.ts
@@ -46,4 +46,28 @@ describe("Extract Component Types", () => {
       }
     `)
   })
+
+  it("should handle arbitrary component names", () => {
+    const theme = {
+      components: {
+        "design-system/Button": {},
+        "design-system_Button": {},
+      },
+    }
+
+    const componentTypes = extractComponentTypes(theme)
+
+    expect(componentTypes).toMatchInlineSnapshot(`
+      Object {
+        "design-system/Button": Object {
+          "sizes": Array [],
+          "variants": Array [],
+        },
+        "design-system_Button": Object {
+          "sizes": Array [],
+          "variants": Array [],
+        },
+      }
+    `)
+  })
 })
